% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_prediction.R
\name{fx_bootPerf}
\alias{fx_bootPerf}
\title{Estimate Bootstrap Distributions}
\usage{
fx_bootPerf(
  modelPerfObj,
  bootObj,
  measures = NULL,
  compute.perf = "within",
  df.iter.out = T,
  qrange = c(0.025, 0.975)
)
}
\description{
Estimate CIs and p-values from bootstrap distributions
}
