% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_prediction.R
\name{fx_modelPerf}
\alias{fx_modelPerf}
\title{xxx}
\usage{
fx_modelPerf(modelOutput, decisionThreshold = 0.5, many = T, perm = F)
}
\arguments{
\item{modelOutput}{aa}

\item{decisionThreshold}{bb}

\item{many}{cc}

\item{perm}{If TRUE, .... Otherwise ...}
}
\value{
Describe output of the function
}
\description{
Short description of the function
}
\examples{
# R code showing how to use the function
1+1

}
