% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_modelResample.R
\name{fx_modelResample}
\alias{fx_modelResample}
\title{Apply Machine Learning Framework}
\usage{
fx_modelResample(
  df0,
  cv.type = NULL,
  covar = NULL,
  voi = NULL,
  outcome = NULL,
  model.type = NULL,
  nresample = 1,
  dthresh = 0.5,
  z.pred = F,
  n.cores = 20,
  balance.col = NULL,
  partitions = NULL
)
}
\arguments{
\item{df0}{data frame including all observations (data frame)}

\item{cv.type}{cross-validation type ('loocv', 'ltocv', 'n-fold', 'numeric') (string)}

\item{covar}{list of df0 column names for "covariate" (not of specific interest) features (string/list)}

\item{voi}{list of df0 column names for variables/features of interest (string/list)}

\item{outcome}{df0 column name for outcome measure to be predicted (string)}

\item{model.type}{machine learning model ('rf', 'logistic', 'regression', 'rf.regression', 'svm') (string)}

\item{nresample}{number of resamples (numeric)}

\item{dthresh}{decision threshold (numeric)}

\item{z.pred}{standardize predictive features (boolean)}

\item{n.cores}{number of cores (parallel processes) (numeric/integer)}

\item{balance.col}{df0 column name used for ensuring balanced columns}

\item{partitions}{pre-defined train/test partitions}
}
\value{
A list of length five, containing the following elements:
\itemize{
\item "perfMetrics" Model performance metrics for each individual fold and "across" and "within".
\cr "across":   sum or mean of metric across folds
\cr "within":   mean of metric across folds
\item "cmat.covar": confusion matrix of covariate model (at "dthresh" decision threshold)

\item "cmat.full":    confusion matrix of full model (at "dthresh" decision threshold)

\item "df.allfolds":  data frame for test-related model predictions

\item "parameters": list of relevant specified parameters
}

A list of length five, containing the following elements:
\itemize{
\item "perfMetrics" Model performance metrics for each individual fold and "across" and "within".
\cr "across":   sum or mean of metric across folds
\cr "within":   mean of metric across folds
\itemize{
\item TP: true positive
\item FP: false positive
\item TN: true negative
\item FN: false negative
\item sens: sensitivity
\item spec: specificity
\item ppv: positive predictive value
\item npv: negative predictive value
\item acc: accuracy
\item auc.ROC: area under the curve of ROC curve
\item optThresh: optimal decision threshold determined from training data
}

\item "cmat.covar": confusion matrix of covariate model (at "dthresh" decision threshold)

\item "cmat.full":    confusion matrix of full model (at "dthresh" decision threshold)

\item "df.allfolds":  data frame for test-related model predictions
\itemize{
\item orig.df.row: row in original data frame for specific observation,
\item fold: fold assignment
\item pred.prob.covar: predicted probability of class membership from covariate model
\item pred.prob.full: predicted probability of class membership from full model
\item pred.class.covar: predicted class from covariate model
\item pred.class.full: predicted class from full model
\item actual.class: actual class membership
}

\item "parameters": list of relevant specified parameters
\itemize{
\item "sample.type": cross-validation sampling procedure
\item "class.levels": class levels
\item "model.type": machine learning model framework
\item "covar": specified covariates
\item "voi": specified variables of interest
\item "outcome": name of class being predicted
\item "formula.covar": formula object for covariate model
\item "formula.full": formula object for full model
\item "data.frame": data frame specified (CURRENTLY NOT CORRECTLY SPECIFIED)
\item "cmat.descrip": key for how to understand confusion matrices ()
\item "negative.class": class assigned to probability = 0
\item "positive.class": class assigned to probability = 1
\item "dthresh": decision threshold
\item "z.pred": whether z-scoring of features is specified
\item "nresample": number of resamples
}
}
}
\description{
Apply machine learning framework to specified dataset
}
\examples{
#### Generate data ####
n <- 100

set.seed(1)
group <- factor(sample(c('MDD','HC'),n,replace=T))
age <- rnorm(n,25,5)
sex <- factor(sample(c('male','female'),n,replace=T))
rand.vals1 <- rnorm(n,0,0.75)
set.seed(2)
rand.vals2 <- rnorm(n,0,0.75)
dd <- data.frame(group = group,
                 age = age,
                 sex = sex,
                 f1 = rand.vals1 + as.numeric(group),
                 f2 = rand.vals2)

#### MODEL EXAMPLE 1 #####
## covariates
covar <- c('age','sex')
## variables of interest
voi <- c('f1','f2')
## class outcome
y <- 'group'

## resamples and permutations
nresample <- 10
nperm <- 10
n.cores <- 1 ## 10

## fit classification model
modelObj <- fx_modelResample(df0 = dd, 
                             cv.type = '5-fold',
                             covar = covar, 
                             voi = voi,  
                             outcome = y,
                             model.type = 'rf',
                             nresample = nresample, 
                             dthresh = 0.5,
                             z.pred = F,
                             balance.col = y,
                             n.cores = n.cores)

## determine overall model performance
modelPerfObj <- fx_modelResamplePerf(modelResampleObj = modelObj)
## permutation testing
permObj <- fx_perm(df0 = dd, modelObj = modelObj, nperm = nperm, n.cores = n.cores)
## determine permutation test performance
permPerfObj <- fx_permPerf(permObj = permObj, modelResamplePerf = modelPerfObj)

## Summary of performance measures based on observed data
modelPerfObj$df.summary
## Outcome metrics for each resample
modelPerfObj$df.iter
## Summary of permutation test outcomes
permPerfObj$df.summary
## Outcome metrics for each permutation
permPerfObj$df.iter
## create roc curve plot
fx_rocPlot(modelObj = modelObj, modelPerfObj = modelPerfObj, permPerfObj = permPerfObj, title.text = 'My Title')
}
