% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_modelResamplePerf.R
\name{fx_modelResamplePerf}
\alias{fx_modelResamplePerf}
\title{Estimates Effects Across Model Variants}
\usage{
fx_modelResamplePerf(
  modelResampleObj,
  measures = NULL,
  compute.perf = "within",
  df.iter.out = T,
  qrange = c(0.025, 0.975)
)
}
\arguments{
\item{modelResampleObj}{???}

\item{measures}{???}

\item{compute.perf}{???}

\item{df.iter.out}{???}

\item{qrange}{???}
}
\description{
Estimates effects across model variants
}
