% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_outFile.R
\name{fx_outFile}
\alias{fx_outFile}
\title{Handles Specified Output Files}
\usage{
fx_outFile(outFile, ext.default = NULL)
}
\arguments{
\item{outFile}{???}

\item{ext.default}{???}
}
\description{
Handles specified output files
}
