% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_partition.R
\name{fx_partition}
\alias{fx_partition}
\title{List of Partitions}
\usage{
fx_partition(df0, type = NULL, nresample = NULL, balance.col = NULL)
}
\arguments{
\item{df0}{data frame including all observations}

\item{type}{cross-validation type}

\item{balance.col}{???}

\item{nresaple}{???}
}
\value{
list of same length as subsamples of train/test partitions.
Each list element contains "test", "train" and "sample.type" variables:
\itemize{
\item "test": is numeric list, values correspond to rows in df0 assigned to model "test" group
\item "train": is numeric list, values correspond to rows in df0 assigned to model "train" group
\item "sample.type": description of cross-validation method (e.g., 5-fold)
}
}
\description{
List of partitions to apply to data frame for CV
}
