% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_partition2.R
\name{fx_partition2}
\alias{fx_partition2}
\title{Extract CV Fold Assignment}
\usage{
fx_partition2(modelObj = NULL)
}
\arguments{
\item{modelObj}{model object, containing observed model performances}
}
\value{
update
\itemize{
\item all.partitions: list of length nresamples (as defined within model Obj input). Each element contains list of length nfolds (i.e., number of folds within each resample). Each element of this sublist is a list containing three variables.
\item train: row indices to be assigned to training dataset
\item test: row indices to be assigned to testing dataset
\item sample.type: cross-validation sampling type (e.g, 5-fold)
}
}
\description{
Extract CV fold assignment from a modelObj
}
