% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_perm.R
\name{fx_perm}
\alias{fx_perm}
\title{Perform Permutation Resampling}
\usage{
fx_perm(
  df0,
  modelObj,
  nperm = 10,
  perm.positions = NULL,
  partitions = NULL,
  n.cores = 20
)
}
\arguments{
\item{df0}{???}

\item{modelObj}{???}

\item{nperm}{???}

\item{perm.positions}{???}

\item{partitions}{???}

\item{n.cores}{???}
}
\description{
Perform permutation resampling to obtain a null distribution
}
