% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_permPerf.R
\name{fx_permPerf}
\alias{fx_permPerf}
\title{Organize Null Distributions}
\usage{
fx_permPerf(
  permObj,
  modelResamplePerf,
  measures = NULL,
  compute.perf = "within",
  df.iter.out = T
)
}
\arguments{
\item{permObj}{???}

\item{modelResamplePerf}{???}

\item{measures}{???}

\item{compute.perf}{???}

\item{df.iter.out}{???}
}
\description{
Estimate p-values, organize null distributions
}
