% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_plot.R
\name{fx_plot}
\alias{fx_plot}
\title{Plot Null/Bootstrap Distributions}
\usage{
fx_plot(perfObj, outFile = NULL)
}
\arguments{
\item{perfObj}{???}

\item{outFile}{???}
}
\description{
Plot observed vs. null/bootstrap distributions
}
