% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_rocCompute.R
\name{fx_rocCompute}
\alias{fx_rocCompute}
\title{Derive AUC ROC}
\usage{
fx_rocCompute(pred.prob, actual.class, class.levels)
}
\arguments{
\item{pred.prob}{???}

\item{actual.class}{???}

\item{class.levels}{???}
}
\value{
list of length five, containing the following elements: roc.auc, optimal.threshold
}
\description{
Derive AUC ROC and optimal threshold values
}
