% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_rocPlot.R
\name{fx_rocPlot}
\alias{fx_rocPlot}
\title{Create AUC-ROC Plots}
\usage{
fx_rocPlot(
  modelObj = NULL,
  modelPerfObj = NULL,
  permPerfObj = NULL,
  title.text = NULL
)
}
\arguments{
\item{modelObj}{model object, containing observed model performances.}

\item{modelPerfObj}{object with summary measures of observed model performance.}

\item{permPerfObj}{object with summary measures of permutation model performance.}

\item{model.type}{string for title of ROC curve plot (string).}
}
\value{
ggplot object of ROC curve
}
\description{
Create AUC-ROC plots
}
