% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_sample.R
\name{fx_sample}
\alias{fx_sample}
\title{Break up Data into Train and Test}
\usage{
fx_sample(df0, partition)
}
\arguments{
\item{df0}{data frame including all observations (data frame).}

\item{partition}{list object containing train and test group assignment.}
}
\value{
A list of length three,containing subsets of original data frame and model parameters
\itemize{
\item df.train: data frame containing only rows assigned to train set
\item df.test: data frame containing only rows assigned to test set
\item parameters: list of relevant specified parameters
\itemize{
\item "sample.type": cross-validation sampling procedure
\item "train.rows": row indices assigned to train set
\item "test.rows": row indices assigned to test set
\item "data.frame": data frame specified (CURRENTLY NOT CORRECTLY SPECIFIED)
\item "nresample": number of resamples
\item "balance.col": column to use for balancing groups
}
}
}
\description{
Break up single data frame into train and test datasets.
}
