% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_scramble.R
\name{fx_scramble}
\alias{fx_scramble}
\title{Scrambled Group Assignment}
\usage{
fx_scramble(df0, outcome, resample = F, outcome.class = T)
}
\arguments{
\item{df0}{???}

\item{outcome}{???}

\item{resample}{???}

\item{outcome.class}{???}
}
\description{
Create data frame with scrambled group assignment
}
