% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_modelPerf.R
\name{fx_modelPerf}
\alias{fx_modelPerf}
\title{Summary Measures of Model Performance}
\usage{
fx_modelPerf(modelOutput, dthresh = 0.5, many = T, perm = F)
}
\arguments{
\item{perm}{}
}
\value{
A list of length five, containing the following elements:
\itemize{
\item "perfMetrics" Model performance metrics for each individual fold and "across" and "within".
\cr "across":   sum or mean of metric across folds
\cr "within":   mean of metric across folds
\itemize{
\item TP: true positive
\item FP: false positive
\item TN: true negative
\item FN: false negative
\item sens: sensitivity
\item spec: specificity
\item ppv: positive predictive value
\item npv: negative predictive value
\item acc: accuracy
\item auc.ROC: area under the curve of ROC curve
\item optThresh: optimal decision threshold determined from training data
}

\item "cmat.covar": confusion matrix of covariate model (at "dthresh" decision threshold)

\item "cmat.full":    confusion matrix of full model (at "dthresh" decision threshold)

\item "df.allfolds":  data frame for test-related model predictions
\itemize{
\item orig.df.row: row in original data frame for specific observation,
\item fold: fold assignment
\item pred.prob.covar: predicted probability of class membership from covariate model
\item pred.prob.full: predicted probability of class membership from full model
\item pred.class.covar: predicted class from covariate model
\item pred.class.full: predicted class from full model
\item actual.class: actual class membership
}

\item "parameters": list of relevant specified parameters
\itemize{
\item "sample.type": cross-validation sampling procedure
\item "class.levels": class levels
\item "model.type": machine learning model framework
\item "covar": specified covariates
\item "voi": specified variables of interest
\item "outcome": name of class being predicted
\item "formula.covar": formula object for covariate model
\item "formula.full": formula object for full model
\item "data.frame": data frame specified (CURRENTLY NOT CORRECTLY SPECIFIED)
\item "cmat.descrip": key for how to understand confusion matrices ()
\item "negative.class": class assigned to probability = 0
\item "positive.class": class assigned to probability = 1
\item "dthresh": decision threshold
\item "z.pred": whether z-scoring of features is specified
\item "nresample": number of resamples
}
}
}
\description{
Derive summary measures of model performance
}
\examples{
## TO BE DONE
}
